#3: Refactor package name in OpenAI API and chat files
