+++ pkg/providers/openai/api.go
@@ -1,4 +1,4 @@
-package main
+package openai
 
 import (
 	"fmt"
